+++ inter/__init__.py
@@ -1,5 +1,6 @@
-from inter.inter import URL, Inter  # noqa
+from inter._client import Client  # noqa
+from inter._inter import Inter, Operation  # noqa
 
 __author__ = """Lucas Rangel Cezimbra"""
 __email__ = 'lucas@cezimbra.tec.br'
-__version__ = '0.0.3'
+__version__ = '0.0.4'