+++ CHANGELOG.md
@@ -1,6 +1,21 @@
 # Changelog
 
 
-## (unreleased)
+## 0.0.4 (2022-10-23)
+- Add High Level API
+- Fix Read the Docs
+- Fix Coveralls
 
-* First release on PyPI.
+
+## 0.0.3 (2022-10-15)
+- Add Inter.get_balance
+- Improve docs
+
+
+## 0.0.2 (2022-10-15)
+- Fix build
+
+
+## 0.0.1 (2022-10-15)
+- First PyPI version
+    * authentication and get_statements